+++ server/model/base/compras.dao.base.php
@@ -7,15 +7,16 @@
   * @access private
   * 
   */
-abstract class ComprasDAOBase
+abstract class ComprasDAOBase extends TablaDAO
 {
 
 	/**
-	  *	metodo save 
+	  *	Guardar registros. 
 	  *	
 	  *	Este metodo guarda el estado actual del objeto {@link Compras} pasado en la base de datos. La llave 
-	  *	primaria indicara que instancia va a ser actualizado en base de datos. Si la llave primara 
-	  *	no esta definicda en el objeto, entonces save() creara una nueva fila.
+	  *	primaria indicara que instancia va a ser actualizado en base de datos. Si la llave primara o combinacion de llaves
+	  *	primarias describen una fila que no se encuentra en la base de datos, entonces save() creara una nueva fila, insertando
+	  *	en ese objeto el ID recien creado.
 	  *	
 	  *	@static
 	  * @param Compras [$compras] El objeto de tipo Compras
@@ -23,11 +24,11 @@ abstract class ComprasDAOBase
 	  **/
 	public static final function save( &$compras )
 	{
-		if(  $compras->getIdCompra()  )
+		if( self::getByPK(  $compras->getIdCompra() ) === NULL )
 		{
-			return ComprasDAOBase::update( $compras) ;
-		}else{
 			return ComprasDAOBase::create( $compras) ;
+		}else{
+			return ComprasDAOBase::update( $compras) ;
 		}
 	}
 
@@ -35,11 +36,11 @@ abstract class ComprasDAOBase
 	/**
 	  *	Obtener {@link Compras} por llave primaria. 
 	  *	
-	  * This will create and load {@link Compras} objects contents from database 
-	  * using given Primary-Key as identifier. 
+	  * Este metodo cargara un objeto {@link Compras} de la base de datos 
+	  * usando sus llaves primarias. 
 	  *	
 	  *	@static
-	  * @return Objeto Un objeto del tipo {@link Compras}.
+	  * @return Objeto Un objeto del tipo {@link Compras}. NULL si no hay tal registro.
 	  **/
 	public static final function getByPK(  $id_compra )
 	{
@@ -47,6 +48,7 @@ abstract class ComprasDAOBase
 		$params = array(  $id_compra );
 		global $db;
 		$rs = $db->GetRow($sql, $params);
+		if(count($rs)==0)return NULL;
 		return new Compras( $rs );
 	}
 
@@ -95,7 +97,7 @@ abstract class ComprasDAOBase
 	  *	  }
 	  * </code>
 	  *	@static
-	  * @param Objeto Un objeto del tipo {@link Compras}.
+	  * @param Compras [$compras] El objeto de tipo Compras
 	  **/
 	public static final function search( $compras )
 	{
@@ -160,7 +162,8 @@ abstract class ComprasDAOBase
 	  * aqui, sin embargo. El valor de retorno indica cuántas filas se vieron afectadas.
 	  *	
 	  * @internal private information for advanced developers only
-	  * @param Objeto El objeto del tipo {@link Compras} a actualizar. 
+	  * @return Filas afectadas
+	  * @param Compras [$compras] El objeto de tipo Compras a actualizar.
 	  **/
 	private static final function update( $compras )
 	{
@@ -176,6 +179,7 @@ abstract class ComprasDAOBase
 			$compras->getIdCompra(), );
 		global $db;
 		$db->Execute($sql, $params);
+		return $db->Affected_Rows();
 	}
 
 
@@ -186,10 +190,11 @@ abstract class ComprasDAOBase
 	  * contenidos del objeto Compras suministrado. Asegurese
 	  * de que los valores para todas las columnas NOT NULL se ha especificado 
 	  * correctamente. Despues del comando INSERT, este metodo asignara la clave 
-	  * primaria generada en el objeto Compras.
+	  * primaria generada en el objeto Compras dentro de la misma transaccion.
 	  *	
 	  * @internal private information for advanced developers only
-	  * @param Objeto El objeto del tipo {@link Compras} a crear. 
+	  * @return Filas afectadas
+	  * @param Compras [$compras] El objeto de tipo Compras a crear.
 	  **/
 	private static final function create( &$compras )
 	{
@@ -205,7 +210,10 @@ abstract class ComprasDAOBase
 		 );
 		global $db;
 		$db->Execute($sql, $params);
+		$ar = $db->Affected_Rows();
+		if($ar == 0) return 0;
 		$compras->setIdCompra( $db->Insert_ID() );
+		return $ar;
 	}
 
 
@@ -214,22 +222,23 @@ abstract class ComprasDAOBase
 	  *	
 	  * Este metodo eliminara la informacion de base de datos identificados por la clave primaria
 	  * en el objeto Compras suministrado. Una vez que se ha suprimido un objeto, este no 
-	  * puede ser restaurado llamando a save(). Restaurarlo solo se puede hacer usando el metodo create(), 
-	  * pero el objeto resultante tendra una diferente clave primaria de la que estaba en el objeto eliminado. 
-	  * Si no puede encontrar eliminar fila coincidente, NotFoundException sera lanzada.
+	  * puede ser restaurado llamando a save(). save() al ver que este es un objeto vacio, creara una nueva fila 
+	  * pero el objeto resultante tendra una clave primaria diferente de la que estaba en el objeto eliminado. 
+	  * Si no puede encontrar eliminar fila coincidente a eliminar, Exception sera lanzada.
 	  *	
-	  * @param Objeto El objeto del tipo {@link Compras} a eliminar. 
+	  *	@throws Exception Se arroja cuando el objeto no tiene definidas sus llaves primarias.
+	  *	@return int El numero de filas afectadas.
+	  * @param Compras [$compras] El objeto de tipo Compras a eliminar
 	  **/
 	public static final function delete( &$compras )
 	{
+		if(self::getByPK($compras->getIdCompra()) === NULL) throw new Exception('Campo no encontrado.');
 		$sql = "DELETE FROM compras WHERE  id_compra = ?;";
-
-		$params = array( 
-			$compras->getIdCompra(), );
-
+		$params = array( $compras->getIdCompra() );
 		global $db;
 
 		$db->Execute($sql, $params);
+		return $db->Affected_Rows();
 	}
